+++ src/frontend/investments.go
@@ -38,8 +38,14 @@ func setInvTableKeybinds(t *updatableTable, inf investmentForm) {
 		} else if event.Rune() == 'd' { // delete investment
 			row, _ := t.GetSelection()
 			id := t.getCellInt(row, 0)
-			backend.DeleteInvestment(id)
-			t.update(t.fGetData(t.curPage))
+			showModal("Confirm delete? (y/n)", func() {
+				backend.DeleteInvestment(id)
+				t.update(t.fGetData(t.curPage))
+				// set focus if deleted last row
+				if row > t.GetRowCount()-1 {
+					t.Select(max(0, row-1), 0)
+				}
+			}, t)
 		} else if event.Rune() == 'e' { // edit investment
 			row, _ := t.GetSelection()
 			id := t.getCellInt(row, 0)
